feat: generate abstract classes for services
